🎨 Imprv(structure): Move store and interfaces to shared
