+++ src/pages/workbench/Workbench.tsx
@@ -1,4 +1,4 @@
-import { CheckboxStatus } from '@/src/_interfaces/_reviewCheckInterfaces';
+import { CheckboxStatus } from '@/src/shared/interfaces/_reviewCheckInterfaces';
 import Button from '@mui/material/Button';
 import ButtonGroup from '@mui/material/ButtonGroup';
 import React, { useState } from 'react';
