+++ src/shared/interfaces/_bookInterfaces.ts
