+++ src/shared/interfaces/_progressInterfaces.ts
