+++ src/shared/interfaces/combinedInterfaces.ts
