+++ src/shared/interfaces/index.ts
