+++ src/shared/store/index.ts
@@ -1,5 +1,5 @@
 import { create } from "zustand";
-import { BoundState } from "../_interfaces";
+import { BoundState } from "../interfaces";
 import createManagementSlice from "./managementStore";
 import createProgressSlice from "./progressStore";
 import createReviewCheckSlice from "./reviewCheckStore";
